Rename rollupServer to arb-validator (and evil version too)
